chore: move API request wrapper classes to appropriate class and delete outdated legacy wrappers
